requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
--extra-index-url https://download.pytorch.org/whl/cpu
azure-functions
pandas
numpy
psycopg2-binary
torch
sysidentpy
sqlalchemy
scikit-learn

